Display basic information about a Maslow need in the form
When Artefact#need is present, display some basic
information about the user need in the artefact
form - the role, goal and benefit.

When Artefact#need returns nil - when the need
doesn't exist, isn't a Maslow need ID, or when
the Need API request fails - continue to show the
need_id form field. (The field should still be
disabled if it's a Maslow need ID as per previous
changes to Panopticon.)

This also includes some tidying up of the
artefact_edit_test.rb integration test so that
all the need tests are contained together using
consistent test data.

Cherry pick conflicts:

	app/views/artefacts/_form.html.erb
	test/integration/artefacts_edit_test.rb