Batch retagging code
To execute, run

    rake tags:batch INPUT OUTPUT

If INPUT provided, will read csv file from location (defaults to `ROOT/data/batch_tag.csv`).
Required format is a column named `slug`. Any columns starting with `keyword` or
unnamed are assumed to be tags.

The OUTPUT file (defaulting to `ROOT/batch_tag.log.csv`) holds the list of files that were
processed. This has a column named error which reports any problems tagging the artefacts.
This is, helpfully, in a format that can be re-read by the batch tagger, for progressive
improvement of the data source. INPUT must be specified if you want to specify a different output.